contract fixes

+++ contracts/contracts/Storage.sol
@@ -72,7 +72,7 @@ contract Storage {
     mapping(uint32 => BlockDEPRECATED) internal blocksDEPRECATED;
 
     /// @notice Flag indicates that a user has exited in the exodus mode certain token balance (per account id and tokenId)
-    mapping(uint32 => mapping(uint16 => bool)) public performedExodus;
+    mapping(uint32 => mapping(uint32 => bool)) public performedExodus;
 
     /// @notice Flag indicates that exodus (mass exit) mode is triggered
     /// @notice Once it was raised, it can not be cleared again, and all users must exit
